Split/refactor school-facing models
Split `src/models/school.js` to put each of the four embedded models in a separate file:
+ `course.js`
+ `department.js`
+ `faculty.js`
+ `record.js`
- `school.js`.
